Models namespace rename
