+++ src/main/scala/com/wordnik/swagger/codegen/BasicAndroidJavaClient.scala
@@ -18,7 +18,7 @@ package com.wordnik.swagger.codegen
 
 import com.wordnik.swagger.codegen.BasicJavaGenerator
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 object BasicAndroidJavaClient extends BasicAndroidJavaGenerator {
   def main(args: Array[String]) = generateClient(args)
