+++ src/main/scala/com/wordnik/swagger/codegen/BasicCSharpGenerator.scala
@@ -16,7 +16,7 @@
 
 package com.wordnik.swagger.codegen
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 object BasicCSharpGenerator extends BasicCSharpGenerator {
   def main(args: Array[String]) = generateClient(args)
