+++ src/main/scala/com/wordnik/swagger/codegen/BasicRubyGenerator.scala
@@ -16,7 +16,7 @@
 
 package com.wordnik.swagger.codegen
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 import java.io.File
