+++ src/main/scala/com/wordnik/swagger/codegen/BasicScalaGenerator.scala
@@ -16,7 +16,7 @@
 
 package com.wordnik.swagger.codegen
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 object BasicScalaGenerator extends BasicScalaGenerator {
   def main(args: Array[String]) = generateClient(args)
