+++ src/main/scala/com/wordnik/swagger/codegen/SpecConverter.scala
@@ -1,7 +1,7 @@
 package com.wordnik.swagger.codegen
 
 import com.wordnik.swagger.codegen.util.{ ResourceExtractor, ApiExtractor }
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 import java.io.File
 
@@ -11,7 +11,7 @@ import org.json4s.jackson.Serialization.{ read, write }
 
 object SpecConverter {
   def main(args: Array[String]) = {
-    implicit val formats = SwaggerSerializers.formats("1.2")
+    implicit val formats = SwaggerModelSerializer.formats("1.2")
 
     if(args == null || args.length < 2) {
       println("Usage: SpecConverter {host} {outputDir}\nIf no API key is required, use an empty string")
