+++ src/main/scala/com/wordnik/swagger/codegen/model/SwaggerModels.scala
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.wordnik.swagger.model
+package com.wordnik.swagger.codegen.model
 
 import scala.collection.mutable.LinkedHashMap
