+++ src/main/scala/com/wordnik/swagger/codegen/spec/SwaggerSpecValidator.scala
@@ -16,7 +16,7 @@
 
 package com.wordnik.swagger.codegen.spec
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 import com.wordnik.swagger.codegen.PathUtil
 import com.wordnik.swagger.codegen.spec.SwaggerSpec._
 import com.wordnik.swagger.codegen.util.CoreUtils
