+++ src/main/scala/com/wordnik/swagger/codegen/util/ApiExtractor.scala
@@ -16,7 +16,7 @@
 
 package com.wordnik.swagger.codegen.util
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 import org.json4s.jackson.JsonMethods._
 import org.json4s.jackson.Serialization.read
@@ -29,7 +29,7 @@ import scala.collection.mutable.{ ListBuffer, HashMap, HashSet }
 
 object ApiExtractor extends RemoteUrl {
   def fetchApiListings(version: String, basePath: String, apis: List[ApiListingReference], authorization: Option[AuthorizationValue] = None): List[ApiListing] = {
-    implicit val formats = SwaggerSerializers.formats(version)
+    implicit val formats = SwaggerModelSerializer.formats(version)
     (for (api <- apis) yield {
       try{
         val json = (basePath.startsWith("http")) match {
@@ -57,7 +57,7 @@ object ApiExtractor extends RemoteUrl {
   }
 
   def extractApiOperations(version: String, basePath: String, references: List[ApiListingReference], authorization: Option[AuthorizationValue] = None) = {
-    implicit val formats = SwaggerSerializers.formats(version)
+    implicit val formats = SwaggerModelSerializer.formats(version)
     for (api <- references) yield {
       val json = basePath.startsWith("http") match {
         case true => {
