+++ src/test/scala/swaggerSpec1_1/UtilsTest.scala
@@ -15,7 +15,7 @@
  */
 package swaggerSpec1_1
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 import com.wordnik.swagger.codegen.util.{ResourceExtractor, ApiExtractor, CoreUtils}
 
 import org.junit.runner.RunWith
