+++ src/test/scala/swaggerSpec1_2/ModelSerializersTest.scala
@@ -1,6 +1,6 @@
 package swaggerSpec1_2
 
-import com.wordnik.swagger.model._
+import com.wordnik.swagger.codegen.model._
 
 import org.json4s._
 import org.json4s.JsonDSL._
@@ -16,7 +16,7 @@ import scala.collection.mutable.LinkedHashMap
 
 @RunWith(classOf[JUnitRunner])
 class ResourceListingSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize an ResourceListing with no apis" in {
     val jsonString = """
@@ -70,7 +70,7 @@ class ResourceListingSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ApiListingReferenceSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize an ApiListingReference" in {
     val jsonString = """
@@ -97,7 +97,7 @@ class ApiListingReferenceSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ApiListingSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize an ApiListing" in {
     val jsonString = """
@@ -141,7 +141,7 @@ class ApiListingSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ApiDescriptionSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
   it should "deserialize an ApiDescription with no ops" in {
     val jsonString = """
     {
@@ -248,7 +248,7 @@ class ApiDescriptionSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class OperationSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize an Operation" in {
     val jsonString = """
@@ -413,7 +413,7 @@ class OperationSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ErrorResponseSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize an Response" in {
     val jsonString = """
@@ -440,7 +440,7 @@ class ErrorResponseSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ParameterSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize another param" in {
     val jsonString = """
@@ -508,7 +508,7 @@ class ParameterSerializersTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ModelSerializationTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize a model" in {
     val jsonString = """
@@ -651,7 +651,7 @@ class ModelSerializationTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ModelRefSerializationTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize a model ref" in {
     val jsonString = """
@@ -678,7 +678,7 @@ class ModelRefSerializationTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class ModelPropertySerializationTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize a model property with allowable values and ref" in {
     val jsonString = """
@@ -829,7 +829,7 @@ class ModelPropertySerializationTest extends FlatSpec with ShouldMatchers {
 
 @RunWith(classOf[JUnitRunner])
 class AllowableValuesSerializersTest extends FlatSpec with ShouldMatchers {
-  implicit val formats = SwaggerSerializers.formats("1.2")
+  implicit val formats = SwaggerModelSerializer.formats("1.2")
 
   it should "deserialize allowable value list" in {
     val allowableValuesListString = """
